requirements.txt
python-dotenv
sse_starlette
sentence-transformers
uvicorn
dashscope
pymupdf
langserve
langchain-community
langchain_openai
langchain-core
langchain_chroma
langgraph
langchain-milvus
chromadb
xinference_client
pyjwt
zhipuai
elasticsearch
jieba
nltk
openai==0.27.6
zhipuai==1.0.7
langchain==0.0.292